+++ package.json
@@ -30,10 +30,10 @@
     "@envelop/graphql-jit": "^1.1.1",
     "@envelop/parser-cache": "^2.2.0",
     "@envelop/validation-cache": "^2.2.0",
-    "@faststore/api": "^1.9.11",
-    "@faststore/graphql-utils": "^1.9.9",
-    "@faststore/sdk": "^1.9.11",
-    "@faststore/ui": "^1.9.11",
+    "@faststore/api": "^1.9.17",
+    "@faststore/graphql-utils": "^1.9.14",
+    "@faststore/sdk": "https://pkg.csb.dev/vtex/faststore/commit/9a30a49c/@faststore/sdk",
+    "@faststore/ui": "^1.9.19",
     "graphql": "^15.0.0",
     "include-media": "^1.4.10",
     "next": "^12.1.6",
@@ -43,11 +43,12 @@
     "react-dom": "^18.2.0",
     "react-intersection-observer": "^8.32.5",
     "sass": "^1.44.0",
-    "swr": "^1.3.0"
+    "swr": "^1.3.0",
+    "yarn": "^1.22.19"
   },
   "devDependencies": {
     "@cypress/code-coverage": "^3.9.10",
-    "@faststore/lighthouse": "^1.9.10",
+    "@faststore/lighthouse": "^1.9.14",
     "@graphql-codegen/cli": "^2.6.2",
     "@graphql-codegen/typescript": "^2.5.1",
     "@graphql-codegen/typescript-operations": "^2.4.2",
