+++ src/components/cart/CartItem/CartItem.tsx
@@ -1,21 +1,22 @@
+import { sendAnalyticsEvent } from '@faststore/sdk'
 import { Card, CardActions, CardContent, CardImage } from '@faststore/ui'
+import { useCallback, useMemo } from 'react'
 import type {
   AddToCartEvent,
   CurrencyCode,
   RemoveFromCartEvent,
 } from '@faststore/sdk'
-import { sendAnalyticsEvent, useSession } from '@faststore/sdk'
-import { useCallback, useMemo } from 'react'
 
 import Button from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
 import { Image } from 'src/components/ui/Image'
 import Price from 'src/components/ui/Price'
 import QuantitySelector from 'src/components/ui/QuantitySelector'
-import { useCart } from 'src/sdk/cart/useCart'
 import { useRemoveButton } from 'src/sdk/cart/useRemoveButton'
 import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
-import type { CartItem as ICartItem } from 'src/sdk/cart/validate'
+import { useSession } from 'src/sdk/session'
+import type { CartItem as ICartItem } from 'src/sdk/cart'
+import { cartStore } from 'src/sdk/cart'
 import type { AnalyticsItem } from 'src/sdk/analytics/types'
 
 import styles from './cart-item.module.scss'
@@ -67,16 +68,15 @@ interface Props {
 
 function CartItem({ item }: Props) {
   const btnProps = useRemoveButton(item)
-  const { updateItemQuantity } = useCart()
   const { sendCartItemEvent } = useCartItemEvent()
 
   const onQuantityChange = useCallback(
     (quantity: number) => {
       sendCartItemEvent(item, quantity)
 
-      updateItemQuantity(item.id, quantity)
+      cartStore.updateItemQuantity(item.id, quantity)
     },
-    [item, sendCartItemEvent, updateItemQuantity]
+    [item, sendCartItemEvent]
   )
 
   return (
