+++ src/components/product/OutOfStock/OutOfStock.stories.mdx
@@ -1,5 +1,5 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs'
-import { SessionProvider } from '@faststore/sdk'
+
 import {
   TokenTable,
   TokenRow,
