+++ src/components/regionalization/RegionalizationInput/RegionalizationInput.tsx
@@ -1,8 +1,7 @@
-import { useSession } from '@faststore/sdk'
 import { useRef, useState } from 'react'
 
 import InputText from 'src/components/ui/InputText'
-import { validateSession } from 'src/sdk/session/validate'
+import { sessionStore, validateSession, useSession } from 'src/sdk/session'
 
 interface Props {
   closeModal: () => void
@@ -10,7 +9,7 @@ interface Props {
 
 function RegionInput({ closeModal }: Props) {
   const inputRef = useRef<HTMLInputElement>(null)
-  const { setSession, isValidating, ...session } = useSession()
+  const session = useSession()
   const [errorMessage, setErrorMessage] = useState<string>('')
 
   const handleSubmit = async () => {
@@ -29,7 +28,7 @@ function RegionInput({ closeModal }: Props) {
       })
 
       if (newSession) {
-        setSession(newSession)
+        sessionStore.set(newSession)
       }
 
       closeModal()
