+++ src/components/search/History/index.tsx
@@ -1,2 +1,1 @@
 export { default as SearchHistory } from './SearchHistory'
-export type { SearchHistoryProps } from './SearchHistory'
