+++ src/components/sections/ProductDetails/ProductDetails.tsx
@@ -1,5 +1,5 @@
-import { sendAnalyticsEvent, useSession } from '@faststore/sdk'
 import { gql } from '@faststore/graphql-utils'
+import { sendAnalyticsEvent } from '@faststore/sdk'
 import { useEffect, useState } from 'react'
 import type { CurrencyCode, ViewItemEvent } from '@faststore/sdk'
 
@@ -14,6 +14,7 @@ import QuantitySelector from 'src/components/ui/QuantitySelector'
 import { useBuyButton } from 'src/sdk/cart/useBuyButton'
 import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
 import { useProduct } from 'src/sdk/product/useProduct'
+import { useSession } from 'src/sdk/session'
 import type { ProductDetailsFragment_ProductFragment } from '@generated/graphql'
 import type { AnalyticsItem } from 'src/sdk/analytics/types'
