+++ src/sdk/analytics/hooks/useViewItemListEvent.ts
@@ -1,7 +1,8 @@
+import { sendAnalyticsEvent } from '@faststore/sdk'
 import { useCallback } from 'react'
 import type { CurrencyCode, ViewItemListEvent } from '@faststore/sdk'
-import { sendAnalyticsEvent, useSession } from '@faststore/sdk'
 
+import { useSession } from 'src/sdk/session'
 import type { ProductSummary_ProductFragment } from '@generated/graphql'
 
 import type { AnalyticsItem } from '../types'
