+++ src/sdk/cart/useBuyButton.ts
@@ -1,15 +1,15 @@
+import { sendAnalyticsEvent } from '@faststore/sdk'
 import { useCallback } from 'react'
-import { sendAnalyticsEvent, useSession } from '@faststore/sdk'
 import type { CurrencyCode, AddToCartEvent } from '@faststore/sdk'
 
+import { cartStore } from 'src/sdk/cart'
+import type { CartItem } from 'src/sdk/cart'
 import type { AnalyticsItem } from 'src/sdk/analytics/types'
-import type { CartItem } from 'src/sdk/cart/validate'
 
+import { useSession } from '../session'
 import { useUI } from '../ui/Provider'
-import { useCart } from './useCart'
 
 export const useBuyButton = (item: CartItem | null) => {
-  const { addItem } = useCart()
   const { openCart } = useUI()
   const {
     currency: { code },
@@ -47,10 +47,10 @@ export const useBuyButton = (item: CartItem | null) => {
         },
       })
 
-      addItem(item)
+      cartStore.addItem(item)
       openCart()
     },
-    [addItem, code, item, openCart]
+    [code, item, openCart]
   )
 
   return {
