+++ src/sdk/product/useProductLink.ts
@@ -1,9 +1,10 @@
-import { sendAnalyticsEvent, useSession } from '@faststore/sdk'
+import { sendAnalyticsEvent } from '@faststore/sdk'
 import { useCallback } from 'react'
 import type { CurrencyCode, SelectItemEvent } from '@faststore/sdk'
 
 import type { ProductSummary_ProductFragment } from '@generated/graphql'
 
+import { useSession } from '../session'
 import type { AnalyticsItem, SearchSelectItemEvent } from '../analytics/types'
 
 export type ProductLinkOptions = {
