+++ src/sdk/search/useSuggestions.ts
@@ -1,4 +1,3 @@
-import { useSession } from '@faststore/sdk'
 import { gql } from '@faststore/graphql-utils'
 
 import { useQuery } from 'src/sdk/graphql/useQuery'
@@ -6,6 +5,8 @@ import type {
   SearchSuggestionsQueryQuery as Query,
   SearchSuggestionsQueryQueryVariables as Variables,
 } from '@generated/graphql'
+
+import { useSession } from '../session'
 
 const MAX_SUGGESTIONS = 5
